requirements.txt
# Core dependencies
binance-connector==3.8.1
pandas-ta==0.4.67b0
pydantic[email]==2.5.2
pyyaml==6.0.1
click==8.1.7
python-telegram-bot>=20.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
openai>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-jose>=3.3.0

# Development and testing
pytest>=8.2.0
pytest-asyncio>=0.24.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
httpx>=0.27.0

# Code quality
black>=24.0.0
ruff>=0.5.0
mypy>=1.10.0
types-PyYAML>=6.0